templates/fastapi_redis/pyproject.toml
[project]
name = "fastapi_redis"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sujay Jayakar", email = "sujayakar@convex.dev"},
]
dependencies = ["fastapi[standard]>=0.115.8", "pyjwt[crypto]>=2.10.1", "passlib[bcrypt]>=1.7.4", "sse-starlette>=2.2.1", "aioredis>=2.0.1"]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
